continue refactoring...
